// src/lib/types.ts
export type CompanyStatus = 'Producer' | 'Developer' | 'Explorer' | 'Royalty' | 'Other';
// ...
export type MetricFormat = 'number' | 'currency' | 'percent' | 'moz' | 'koz';